backend/requirements.txt
# -------------------------
# Backend Framework
# -------------------------
fastapi==0.110.3
uvicorn==0.29.0
starlette==0.37.2
websockets==12.0
anyio==4.3.0

# -------------------------
# Scheduling & PubSub
# -------------------------
APScheduler==3.10.4
redis==5.0.4

# -------------------------
# Audio Processing
# -------------------------
numpy==1.26.4
soundfile==0.12.1

# -------------------------
# PyTorch (CUDA 11.8 – SAFE)
# -------------------------
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0

# -------------------------
# Diarization
# -------------------------
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.0

# -------------------------
# Speech Recognition
# -------------------------
openai-whisper==20231117

# -------------------------
# Utilities
# -------------------------
pydantic==2.7.1
python-dotenv==1.0.1
tqdm==4.66.4

# -------------------------
# Embeddings / Vector DB
# -------------------------
sentence-transformers==2.2.2
faiss-cpu==1.7.4

